ppc/xive: Introduce a xive_os_cam_decode() helper
The OS CAM line has a special encoding exploited by the HW. Provide a
helper routine to hide the details to the TIMA command handlers. This
also clarifies the endianness of different variables : 'qw1w2' is
big-endian and 'cam' is native.

Also add a check on the V bit in the TM_PULL_POOL_CTX special
operation.

Signed-off-by: Cédric Le Goater <clg@kaod.org>